requirements.txt
beautifulsoup4
selenium
jsbeautifier
selenium-wire
setuptools
python-dotenv
